+++ config/config.exs
@@ -1,0 +1,3 @@
+import Config
+
+config :rustler_precompiled, :force_build, umya_spreadsheet_ex: true